+++ archunit/src/main/java/com/tngtech/archunit/core/domain/AnnotationProxy.java
@@ -29,9 +29,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Maps;
+import com.tngtech.archunit.base.MayResolveTypesViaReflection;
 import com.tngtech.archunit.base.Optional;
 import com.tngtech.archunit.core.InitialConfiguration;
-import com.tngtech.archunit.core.MayResolveTypesViaReflection;
 
 import static com.google.common.base.Preconditions.checkArgument;
